debug goggle (#167)
- add `all` install target
- drop the goggle mark for tests and the checks from the plugin.

torch-sparse requires compiling. `pip` is not aware of local computer architectures, so you need to compile when installing with pip. `conda install` can handle the computer arch, but you need to be sure you match the torch version, otherwise it will crash.

For now, pip compiles the library everytime. However, the github actions are caching the `pip` stage, so it won't take forever everytime.