[WIP] 58 add goggle (#162)
* First goggle implementation, no cond

* cleanup

* Adding tests and some debugging

* Debugging

* Enabled reproducible results

* getting tests to pass

* Clean up

* Added more argument validation

* Add docstrings

* remove cond test

* Expose node_dim as parameter.

* Added dependencies to setup.cfg

* move goggle reqs to extra

* Swap airfoil dataset for diabetes in goggle tests

* Debugging dependencies

* debug goggle (#167)

- add `all` install target
- drop the goggle mark for tests and the checks from the plugin.

torch-sparse requires compiling. `pip` is not aware of local computer architectures, so you need to compile when installing with pip. `conda install` can handle the computer arch, but you need to be sure you match the torch version, otherwise it will crash.

For now, pip compiles the library everytime. However, the github actions are caching the `pip` stage, so it won't take forever everytime.

* Update README.md

---------

Co-authored-by: Bogdan Cebere <bogdan.cebere@gmail.com>